+++ scripts/module.js
@@ -101,8 +101,13 @@ Hooks.once("init", function() {
     Settings
   };
 
+  // Add render flags to help with autotargeting and elevation of the template.
   CONFIG.MeasuredTemplate.objectClass.RENDER_FLAGS.retarget = {};
   CONFIG.MeasuredTemplate.objectClass.RENDER_FLAGS.refreshPosition.propagate = ["retarget", "refreshShape"];
+
+  CONFIG.MeasuredTemplate.objectClass.RENDER_FLAGS.refreshElevation = { propagate: ["retarget", "refreshShape"]};
+  CONFIG.MeasuredTemplate.objectClass.RENDER_FLAGS.refreshPosition.propagate = ["retarget", "refreshShape", "refreshElevation"];
+
 
   // Tell modules that the module is set up
   Hooks.callAll(`${MODULE_ID}Ready`);
@@ -172,9 +177,6 @@ Hooks.once("ready", async function() {
   if ( promises.length ) await Promise.all(promises);
 
   // Ensure autotargeting is registered on setup.
-  const atMenuChoice = Settings.get(KEYS.AUTOTARGET.MENU);
-  const enabled = Settings.get(KEYS.AUTOTARGET.ENABLED) || atMenuChoice === KEYS.AUTOTARGET.CHOICES.YES;
-  Settings.set(KEYS.AUTOTARGET.ENABLED, enabled);
   registerAutotargeting();
 
   log("Refreshing templates on ready hook.");
@@ -185,6 +187,10 @@ Hooks.once("ready", async function() {
       refreshShape: true
     });
   });
+
+  log("Refreshing autotargeting.");
+  Settings.refreshAutotargeting();
+
 });
 
 Hooks.on("getSceneControlButtons", controls => {
@@ -196,7 +202,7 @@ Hooks.on("getSceneControlButtons", controls => {
     name: "autotarget",
     title: game.i18n.localize("walledtemplates.controls.autotarget.Title"),
     toggle: true,
-    visible: opt === AUTOTARGET.CHOICES.TOGGLE_OFF || opt === AUTOTARGET.CHOICES.TOGGLE_ON,
+    visible: opt === AUTOTARGET.CHOICES.TOGGLE,
     active: Settings.get(AUTOTARGET.ENABLED),
     onClick: toggle => { // eslint-disable-line no-unused-vars
       Settings.toggle(AUTOTARGET.ENABLED);