+++ Changelog.md
@@ -1,3 +1,16 @@
+## 0.7.4
+
+### Features
+Add flag and MeasuredTemplate config to disable autotargeting for specific template.
+Add keybinding to unhide template highlighting and borders. Add toggle to template configuration to force display of specific template highlighting and borders (flags: `forceBorder` and `forceHighlighting`). Closes #93, #95.
+Add toggle to template configuration to disable autotargeting for specific template (flag: `noAutotarget`).
+Add elevation tooltip.
+
+### Bug fixes
+Incorporate PR #89 and #90 from @strongpauly. Thanks!
+Switch to a single toggle option (instead of 2) for autotargeting in settings. Fix for template target sets being shared between targets. Templates now do not autotarget on load if autotargeting is not enabled or the autotarget toggle is disabled. Closes #88.
+Update lib geometry to v0.2.19.
+
 ## 0.7.3
 Force the template to stick to its set distance when the token drag is completed. Ensure drag method is awaiting the underlying method to avoid ghosting when using Elevation Ruler to drag tokens. Closes issue #86.